Complete heap implementation
One thing this excursion taught me was the reason dynamic languages
tend to be object oriented. With implicit assertions like these there
is no mechanism for validating implicit representations. I'm planning
on moving this to a completely flat definition, independent of the
array, and anyone can pass in the wrong array, causing the issue.

After this I'll probably drop Python for a little bit and move on to
ocaml :)